# backend/src/device/device.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  Post,
  Query,
  Req,
  UseGuards,
} from '@nestjs/common';
import {
  ActionType,
  DeviceConfigOverride,
  DeviceFirmwareOverride,
} from '@prisma/client';
import { Request } from 'express';
import { RequirePermission } from '../common/decorators/require-permission.decorator';
import { JwtAuthGuard, JwtPayload } from '../common/guards/jwt-auth.guard';
import { PermissionGuard } from '../common/guards/permission.guard';
import type { ConfigApplyResult } from './config-applier';
import type { DeviceConnectionTestResult } from './device-connection-tester';
import { ApplyConfigDto } from './dto/apply-config.dto';
import { ConfirmFirmwareInstallDto } from './dto/confirm-firmware-install.dto';
import { DeviceConfigOverrideDto } from './dto/device-config-override.dto';
import { DeviceFirmwareOverrideDto } from './dto/device-firmware-override.dto';
import { QueryDeviceDto } from './dto/query-device.dto';
import { RegisterDeviceDto } from './dto/register-device.dto';
import { SimulateConfigOnDeviceDto } from './dto/simulate-config-on-device.dto';
import type {
  ActingUser,
  ConfigWithDeviceOverride,
  ConfirmFirmwareInstallResult,
  DeviceStatusResult,
  DeviceWithCustomer,
  RegisterDeviceResult,
} from './device.service';
import { DeviceService } from './device.service';
import type { DeviceSimulateConfigResult } from './simulate-config-result';

/** Request ที่ผ่าน JwtAuthGuard จะมี user อยู่เสมอ — mirror config.controller.ts */
type AuthenticatedRequest = Request & { user: JwtPayload };

function toActor(req: AuthenticatedRequest): ActingUser {
  return { id: req.user.sub, role: req.user.role };
}

// Device module:
//   POST /devices                 — ลงทะเบียนอุปกรณ์ใหม่ (issue #157 PR 1) ·
//                                 Admin/SuperAdmin เท่านั้น (staff ลงทะเบียน
//                                 ไม่ใช่อุปกรณ์เรียกเอง — endpoint ที่อุปกรณ์
//                                 เรียกเองผ่าน DeviceApiKeyGuard มาใน PR 2/3)
//   POST /devices/:deviceId/rotate-key — ออก/หมุนเวียน key ให้เครื่องที่มี
//                                 อยู่แล้ว (เครื่องเก่าก่อนฟีเจอร์นี้ หรือ
//                                 หมุนเวียน key ที่สงสัยว่ารั่ว) · Admin/
//                                 SuperAdmin เท่านั้นเหมือนกัน
//   GET  /devices                 — Device Search (list + filter)  · ทุก Role
//   GET  /devices/:deviceId        — Device Detail (1 เครื่อง)      · ทุก Role
//   GET  /devices/:deviceId/status — สถานะย่อ configStatus/firmwareStatus
//                                 (issue #245) · ทุก Role — ยังไม่มี online/
//                                 offline (ไม่มี concept check-in ในระบบเลย)
//   POST /devices/:deviceId/test-connection | apply-config | simulate-config
//                                 — ช่างหน้างาน ST/OT ผ่าน Mobile
//   GET  /devices/:deviceId/config — Config ปัจจุบันของอุปกรณ์ (issue #211,
//                                 merge override ที่ approved แล้วจาก #223
//                                 ถ้ามี + pendingOverride ถ้ามีคำขอค้างอยู่)
//                                 — ช่างหน้างาน ST/OT ผ่าน Mobile
//   POST /devices/:deviceId/confirm-firmware-install — ยืนยันติดตั้ง Firmware
//                                 (issue #181) — ช่างหน้างาน ST/OT ผ่าน Mobile
//   POST /devices/:deviceId/config-override — Per-device Config Override —
//                                 ส่งคำขอ (สถานะ pending) (issue #223) —
//                                 ช่างหน้างาน ST เท่านั้น
//
// คิวอนุมัติของ Operation อยู่คนละ controller (`device-config-override.controller.ts`,
// path `/device-config-overrides`, ไม่ nest ใต้ `/devices/:deviceId`) — ดูที่นั่น
//
// **ยังไม่ implement `GET /devices/:deviceId/status`** — มีแต่ spec ใน
// openapi.yaml (schema `DeviceStatus`) ยังไม่เคยมีโค้ดจริง · การคำนวณ
// configStatus/firmwareStatus ต้องออกแบบใหม่ทั้งก้อน (Device ไม่มี FK ตรงไป
// Config/Firmware — ตกลงกับ B บน PR #52) เป็น PR แยกในอนาคต · Device Detail
// (`GET /devices/:deviceId`) รอบนี้คืนแค่ field ของ record Device เอง
@UseGuards(JwtAuthGuard, PermissionGuard)
@Controller('devices')
export class DeviceController {
  constructor(private readonly deviceService: DeviceService) {}

  // issue #157 PR 1 — ลงทะเบียนอุปกรณ์ใหม่ resource `device-registration`
  // action Create · grant ให้ Admin เท่านั้น (SuperAdmin inherit อัตโนมัติผ่าน
  // loop ใน prisma/seed.ts) — mirror ขอบเขตเดียวกับ `device-model` Create/Update
  @Post()
  @RequirePermission('device-registration', ActionType.Create)
  register(
    @Body() dto: RegisterDeviceDto,
    @Req() req: AuthenticatedRequest,
  ): Promise<RegisterDeviceResult> {
    return this.deviceService.register(dto, toActor(req));
  }

  // issue #157 PR 1 (เพิ่มระหว่างทำ) — ออก/หมุนเวียน key ให้เครื่องที่มีอยู่
  // แล้ว (เครื่องเก่าที่ลงทะเบียนก่อนฟีเจอร์นี้ หรือหมุนเวียน key ที่สงสัยว่ารั่ว)
  // resource เดียวกับ register แต่ action Update (แก้ไข ไม่ใช่สร้างใหม่)
  @Post(':deviceId/rotate-key')
  @RequirePermission('device-registration', ActionType.Update)
  @HttpCode(HttpStatus.OK)
  rotateKey(
    @Param('deviceId') deviceId: string,
    @Req() req: AuthenticatedRequest,
  ): Promise<RegisterDeviceResult> {
    return this.deviceService.rotateKey(deviceId, toActor(req));
  }

  // Device Search / Device Detail (Sprint 2 #11) — resource `devices` action
  // Read · RBAC_Matrix.md §2 แถว "Device Search / Device Detail" = R ทุก Role
  // · grant `devices` Read seed ให้ทุก role อยู่แล้ว (prisma/seed.ts — เดิม
  // เตรียมไว้ให้ getDeviceStatus ที่ยังไม่ implement) ไม่ต้อง seed เพิ่ม
  //
  // key ด้วย `Device.deviceId` (เลขเครื่องจริง) ไม่ใช่ `Device.id` (UUID
  // ภายใน) — เหมือน endpoint ช่างหน้างานด้านล่าง
  @Get()
  @RequirePermission('devices', ActionType.Read)
  findAll(@Query() query: QueryDeviceDto): Promise<DeviceWithCustomer[]> {
    return this.deviceService.findAll(query);
  }

  @Get(':deviceId')
  @RequirePermission('devices', ActionType.Read)
  findOne(@Param('deviceId') deviceId: string): Promise<DeviceWithCustomer> {
    return this.deviceService.findByDeviceId(deviceId);
  }

  // issue #245 — เวอร์ชันย่อ: มีแค่ configStatus/firmwareStatus (คำนวณจาก
  // CampaignRolloutTarget ล่าสุด) ยังไม่มี online/offline/lastCheckIn เวลาจริง
  // (ยังไม่มี concept check-in ในระบบเลย — เปิดเป็นคำถามแยกต่างหาก ไม่บล็อก
  // ส่วนนี้) resource/action เดียวกับ Device Detail (ทุก Role ที่ login แล้ว
  // ตามที่ยืนยันกับ B แล้ว) ไม่ต้อง seed grant เพิ่ม
  @Get(':deviceId/status')
  @RequirePermission('devices', ActionType.Read)
  getStatus(@Param('deviceId') deviceId: string): Promise<DeviceStatusResult> {
    return this.deviceService.getStatus(deviceId);
  }

  // resource `device-connection-test` action Read — grant ให้ ST/OT เท่านั้น
  // (ดู prisma/seed.ts) A ยืนยันบน PR #52 ว่ายังไม่เปิดให้ ConfigEngineer/
  // FirmwareEngineer/Operation เพราะยังไม่เห็น use case ชัดเจน
  @Post(':deviceId/test-connection')
  @RequirePermission('device-connection-test', ActionType.Read)
  @HttpCode(HttpStatus.OK)
  testConnection(
    @Param('deviceId') deviceId: string,
  ): Promise<DeviceConnectionTestResult> {
    return this.deviceService.testConnection(deviceId);
  }

  // resource `device-config-apply` action Read — grant ให้ ST/OT เท่านั้น
  // (mirror `device-connection-test` — คนหน้างานที่ใช้ Mobile) ConfigEngineer/
  // FirmwareEngineer/Operation ยังไม่เปิด (ทำ Config บน Web ไม่ได้ apply หน้างาน)
  //
  // action ใช้ `Read` (ไม่ใช่ Update/Create) โดยตั้งใจ — apply-config ไม่ persist
  // อะไรในระบบเรา (fire-and-forget mock, ยืนยัน scope กับ B) จึงไม่ตรงกับ
  // Update/Create ที่หมายถึงแก้ข้อมูลใน DB — pattern เดียวกับ `config-simulation`
  @Post(':deviceId/apply-config')
  @RequirePermission('device-config-apply', ActionType.Read)
  @HttpCode(HttpStatus.OK)
  applyConfig(
    @Param('deviceId') deviceId: string,
    @Body() dto: ApplyConfigDto,
    @Req() req: AuthenticatedRequest,
  ): Promise<ConfigApplyResult> {
    return this.deviceService.applyConfig(deviceId, dto.configId, toActor(req));
  }

  // resource `device-connection-test` action Read — reuse permission เดิม
  // (ST/OT เท่านั้น) endpoint นี้เช็คสัญญาณอุปกรณ์จริงเหมือนกัน ตกลงกับ
  // kittiphong (B) บน PR #92 ว่าใช้สิทธิ์เดียวกัน ไม่ต้อง seed เพิ่ม
  //
  // action `Read` — ไม่ persist อะไร (dry-run readiness check) pattern เดียว
  // กับ `test-connection` / `apply-config`
  @Post(':deviceId/simulate-config')
  @RequirePermission('device-connection-test', ActionType.Read)
  @HttpCode(HttpStatus.OK)
  simulateConfig(
    @Param('deviceId') deviceId: string,
    @Body() dto: SimulateConfigOnDeviceDto,
  ): Promise<DeviceSimulateConfigResult> {
    return this.deviceService.simulateConfig(deviceId, dto.configId);
  }

  // resource `device-current-config` action Read — grant ให้ ST/OT เท่านั้น
  // (mirror device-connection-test/device-config-apply) Config Override
  // Phase 2 (Mobile, issue #211) — ช่างต้องรู้ Config ปัจจุบันของอุปกรณ์ก่อน
  // เปิดหน้า Override
  @Get(':deviceId/config')
  @RequirePermission('device-current-config', ActionType.Read)
  getCurrentConfig(
    @Param('deviceId') deviceId: string,
  ): Promise<ConfigWithDeviceOverride> {
    return this.deviceService.getCurrentConfig(deviceId);
  }

  // resource ใหม่ `device-config-override` action `Override` — grant ST
  // เท่านั้น (mirror `config-override` เดิม, issue #185 — seed.ts) OT ไม่มี
  // สิทธิ์เลยเหมือนกัน Per-device Config Override (issue #223, มติ
  // 2026-09-24) — ดู comment เหนือ `DeviceService.overrideDeviceConfig()` —
  // คืนคำขอที่เพิ่งสร้าง (สถานะ `pending` เสมอ) ไม่ใช่ Config ที่ merge แล้ว
  // เหมือนเดิม เพราะยังไม่มีผลจนกว่า Operation จะอนุมัติ
  @Post(':deviceId/config-override')
  @RequirePermission('device-config-override', ActionType.Override)
  @HttpCode(HttpStatus.OK)
  overrideDeviceConfig(
    @Param('deviceId') deviceId: string,
    @Body() dto: DeviceConfigOverrideDto,
    @Req() req: AuthenticatedRequest,
  ): Promise<DeviceConfigOverride> {
    return this.deviceService.overrideDeviceConfig(deviceId, dto, toActor(req));
  }

  // resource ใหม่ `device-firmware-confirm` action `Create` — grant ให้ ST/OT
  // เท่านั้น (mirror `device-config-apply`) **action `Create` ไม่ใช่ `Read`**
  // ต่างจาก apply-config/test-connection/simulate-config โดยตั้งใจ —
  // endpoint นี้เขียน AuditLog จริง (ไม่ใช่ fire-and-forget ที่ไม่ persist
  // อะไรเลยแบบพวกนั้น) ดู comment เหนือ `DeviceService.confirmFirmwareInstall()`
  @Post(':deviceId/confirm-firmware-install')
  @RequirePermission('device-firmware-confirm', ActionType.Create)
  @HttpCode(HttpStatus.OK)
  confirmFirmwareInstall(
    @Param('deviceId') deviceId: string,
    @Body() dto: ConfirmFirmwareInstallDto,
    @Req() req: AuthenticatedRequest,
  ): Promise<ConfirmFirmwareInstallResult> {
    return this.deviceService.confirmFirmwareInstall(
      deviceId,
      dto,
      toActor(req),
    );
  }

  // resource ใหม่ `device-firmware-override` action `Override` — grant ST
  // เท่านั้น (mirror `device-config-override` — OT ไม่มีสิทธิ์เลยเหมือนกัน)
  // Firmware Override รายเครื่อง (Sprint 3 แถวที่ 24) — ดู comment เหนือ
  // `DeviceService.overrideDeviceFirmware()` — คืนคำขอที่เพิ่งสร้าง (สถานะ
  // `pending` เสมอ) ไม่มีผลกับ `confirmFirmwareInstall()` จนกว่า Operation
  // จะอนุมัติ
  @Post(':deviceId/firmware-override')
  @RequirePermission('device-firmware-override', ActionType.Override)
  @HttpCode(HttpStatus.OK)
  overrideDeviceFirmware(
    @Param('deviceId') deviceId: string,
    @Body() dto: DeviceFirmwareOverrideDto,
    @Req() req: AuthenticatedRequest,
  ): Promise<DeviceFirmwareOverride> {
    return this.deviceService.overrideDeviceFirmware(
      deviceId,
      dto,
      toActor(req),
    );
  }
}