# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GPS Config & Firmware Center API
  description: >
    โครง API เริ่มต้นสำหรับ Sprint 0 — ยังไม่ครบทุก Endpoint
    เพิ่มเติมทีละ Phase ตามที่ Build จริง (ดู ../planning/01_GPS_Build_Reference.md Section 3)
  version: "0.1.0"
servers:
  - url: http://localhost:3001/api/v1
    description: Local dev

# บังคับ Bearer JWT ทุก endpoint โดย default — override เป็น security: [] เฉพาะ endpoint ที่เป็น public
security:
  - bearerAuth: []
...
paths:
...
  /config/{configId}/simulate:
    post:
      tags: [config]
      operationId: simulateConfig
      summary: ทดสอบ Config กับ Device Simulator (ใช้ทั้งฝั่ง Web และ Mobile — v3.5)
      parameters:
        - in: path
          name: configId
          required: true
          schema:
            type: string
      requestBody:
        # required: false (v3.9 — แก้ตาม code review ของ kittiphong บน PR
        # Stage 3: ConfigController.simulate ไม่รับ @Body() เลยจริงๆ — เดิม
        # required: true ทำให้ client ที่ gen จาก spec ตรงๆ ถูกบังคับส่ง body
        # ที่ backend ไม่สนใจอยู่ดี ส่วน `deviceModel` ใน schema ก็ไม่ถูกใช้
        # เลย (ConfigService.simulate ใช้ deviceModel/protocol/fields จาก DB
        # ของ Config นั้นเสมอ ไม่รับจาก client) เก็บ property นี้ไว้เป็นข้อมูล
        # อ้างอิงเฉยๆ
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceModel:
                  type: string
      responses:
        "200":
          description: ผลทดสอบ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
        "400":
          description: deviceModel ไม่ถูกต้องหรือไม่รองรับ
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ทดสอบ Config นี้ — เฉพาะ Role ConfigEngineer, Operation, ST, OT
        "404":
          description: ไม่พบ Config นี้
        "409":
          description: สถานะ Config ปัจจุบันไม่รองรับการทดสอบ (เช่น approved/synced ไปแล้ว)
...
  /devices:
    post:
      tags: [device]
      operationId: registerDevice
      summary: ลงทะเบียนอุปกรณ์ใหม่เข้าระบบ (v3.38, issue #157 PR 1)
      description: >-
        **endpoint ฝั่ง staff** (RBAC_Matrix.md §4.1 — resource ใหม่
        `device-registration` action `Create`, Admin/SuperAdmin เท่านั้น)
        **ไม่ใช่ endpoint ที่อุปกรณ์เรียกเอง** — อุปกรณ์ยังไม่มี API key จนกว่า
        การลงทะเบียนนี้จะสำเร็จ (endpoint ที่อุปกรณ์เรียกเองด้วย key นี้ผ่าน
        guard แยกมาใน PR ถัดไป ดู docs/14_Device_Sync_Proposal.md) · รับ
        `modelId` ไม่รับ `deviceModel` string ตรงๆ — server resolve
        `deviceModel` เองจาก `DeviceModel.name` เสมอ (issue #209) · `protocol`
        ต้องอยู่ใน `DeviceModel.supportedProtocols` ของรุ่นนั้น · `apiKey` ใน
        response เป็นค่าจริง (ไม่ hash) โชว์ได้**ครั้งเดียว**ตอนนี้เท่านั้น —
        ไม่มีทาง GET กลับมาดูซ้ำได้อีก (เก็บเฉพาะ hash ไว้ฝั่ง backend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: ลงทะเบียนสำเร็จ — เก็บ apiKey ที่ได้รับไว้ให้ดี ดูซ้ำไม่ได้แล้ว
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterDeviceResponse"
        "400":
          description: >-
            field บังคับขาด/ผิดรูปแบบ หรือ protocol ที่ระบุไม่อยู่ใน
            supportedProtocols ของ modelId นั้น
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ลงทะเบียนอุปกรณ์ — เฉพาะ Role Admin, SuperAdmin
        "404":
          description: ไม่พบ DeviceModel (modelId) นี้ในทะเบียน
        "409":
          description: มี Device deviceId นี้อยู่แล้วในระบบ

    get:
      tags: [device]
      operationId: listDevices
      summary: ค้นหา/รายการอุปกรณ์ (Device Search — v3.17)
      description: >-
        ทุก Role ที่ login แล้วเรียกได้ (RBAC_Matrix.md §2 แถว "Device Search /
        Device Detail" = R ทุก Role) · filter ทุกตัว optional — ไม่ส่งอะไรมา =
        คืนทุกเครื่อง · `search` match ทั้ง `deviceId` และ `simNumber`
        (contains, case-insensitive) · ยังไม่มี paging (จำนวน Device ใน MVP น้อย)
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
            maxLength: 100
          description: คำค้น — match deviceId หรือ simNumber แบบ contains
        - in: query
          name: deviceModel
          required: false
          schema:
            type: string
            maxLength: 50
        - in: query
          name: protocol
          required: false
          schema:
            type: string
            maxLength: 50
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [registered, installed, decommissioned]
        - in: query
          name: customerId
          required: false
          schema:
            type: string
            format: uuid
          description: >-
            กรองเฉพาะอุปกรณ์ของลูกค้ารายนี้ (v3.28, issue #204) — เครื่องที่
            `customerId` เป็น `null` (ยังไม่ผูกบริษัท) จะไม่ถูกคืนเมื่อส่ง
            filter นี้มา
      responses:
        "200":
          description: รายการ Device (เรียงตาม deviceId)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "400":
          description: >-
            ค่า status ไม่อยู่ใน enum, query param ยาวเกินกำหนด, หรือ
            customerId ไม่ใช่ uuid
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
  /devices/{deviceId}:
    get:
      tags: [device]
      operationId: getDevice
      summary: ดูรายละเอียดอุปกรณ์ 1 เครื่อง (Device Detail — v3.17)
      description: >-
        อ้างด้วย `Device.deviceId` (เลขเครื่องจริง) ไม่ใช่ `Device.id` (UUID
        ภายใน) · คืนเฉพาะ field ของ record Device เอง — สถานะ Config/Firmware
        ของกล่อง (`GET /devices/{deviceId}/status`) ยังไม่ implement
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ข้อมูล Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "404":
          description: ไม่พบ Device (deviceId) นี้
  /devices/{deviceId}/rotate-key:
    post:
      tags: [device]
      operationId: rotateDeviceApiKey
      summary: ออก/หมุนเวียน API key ให้เครื่องที่มีอยู่แล้ว (v3.40, issue #157)
      description: >-
        **endpoint ฝั่ง staff** (resource เดียวกับ `registerDevice` —
        `device-registration` — action `Update` เพราะแก้ไขเครื่องที่มีอยู่แล้ว
        Admin/SuperAdmin เท่านั้น) ใช้ได้ 2 กรณี: (1) ออก key ครั้งแรกให้เครื่อง
        เก่าที่ลงทะเบียนไว้ก่อนฟีเจอร์นี้ (`apiKeyHash` เดิมเป็น `null`) (2)
        หมุนเวียน key เดิมถ้าสงสัยว่ารั่ว — key เก่าใช้ต่อไม่ได้ทันทีที่เรียก
        สำเร็จ ไม่มี grace period · `apiKey` ใน response เป็นค่าจริงโชว์ได้
        **ครั้งเดียว** เหมือน `registerDevice` · **การแจกจ่าย key ใหม่ให้กล่อง
        จริงที่ติดตั้งอยู่แล้วเป็นขอบเขตพี่เลี้ยง** (ระบบเป็น PULL model ไม่มี
        ช่องทาง push เข้าไปหากล่องเลย) endpoint นี้แค่รับประกันว่ามี key ที่
        ถูกต้องรออยู่ให้ไปติดตั้ง
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ออก/หมุนเวียน key สำเร็จ — เก็บ apiKey ที่ได้รับไว้ให้ดี ดูซ้ำไม่ได้แล้ว
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterDeviceResponse"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ออก/หมุนเวียน API key ของอุปกรณ์ — เฉพาะ Role Admin, SuperAdmin
        "404":
          description: ไม่พบ Device (deviceId) นี้
  /devices/{deviceId}/status:
    get:
      tags: [device-status]
      operationId: getDeviceStatus
      summary: ดูสถานะ Config/Firmware ของอุปกรณ์ — เวอร์ชันย่อ (v3.39, issue #245)
      description: >-
        ทุก Role ที่ login แล้วเรียกได้ (resource `devices` action `Read`
        เดียวกับ Device Detail) — `configStatus`/`firmwareStatus` คำนวณจาก
        `CampaignRolloutTarget` ล่าสุดของอุปกรณ์นี้แยกตาม payload type
        (`unknown` = ไม่เคยอยู่ใน rollout ไหนเลย เป็นค่าปกติ ไม่ใช่ error) —
        **ยังไม่มี online/offline หรือเวลา check-in จริง** (ระบบยังไม่มี
        concept ให้อุปกรณ์ check-in เลย — เปิดเป็นคำถามแยกต่างหากอยู่)
        `lastCheckInMessage` จึงเป็น `null` เสมอตอนนี้
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: สถานะล่าสุด
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceStatus"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "404":
          description: ไม่พบ Device นี้
  /devices/{deviceId}/test-connection:
    post:
      tags: [device]
      operationId: testDeviceConnection
      summary: ทดสอบการเชื่อมต่อ/สัญญาณของอุปกรณ์ที่ติดตั้งจริง (v3.11)
      description: >-
        สำหรับช่างหน้างาน (ST/OT ผ่าน Mobile) ทดสอบสัญญาณอุปกรณ์เครื่องที่
        ติดตั้งจริงแล้ว — ต่างจาก `simulateConfig` ที่เป็น dry-run ตรวจค่า field
        ของ Config template (สถานะ `draft`/`testing`). อ้างด้วยเลขเครื่องจริง
        (`Device.deviceId`) ไม่มี request body — server ดึงข้อมูล Device จาก DB เอง
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ผลทดสอบการเชื่อมต่อ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceConnectionTestResult"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ทดสอบการเชื่อมต่ออุปกรณ์ — เฉพาะ Role ST, OT
        "404":
          description: ไม่พบ Device (deviceId) นี้
        "409":
          description: Device ยังไม่ถึงสถานะ installed (ยังเป็น registered หรือ decommissioned ไปแล้ว) จึงทดสอบสัญญาณไม่ได้
  /devices/{deviceId}/apply-config:
    post:
      tags: [device]
      operationId: applyConfigToDevice
      summary: ใส่ Config ที่อนุมัติแล้วให้อุปกรณ์ที่ติดตั้งจริง (v3.13)
      description: >-
        สำหรับช่างหน้างาน (ST/OT ผ่าน Mobile) — เลือก Config ที่ผูกกับ Task
        ที่กำลังทำ แล้วส่งให้อุปกรณ์เครื่องนั้น. **ไม่ persist / ไม่แตะสถานะใดๆ**
        (fire-and-forget) — กล่องจะรับค่าเมื่อเปิดเครื่องครั้งถัดไป
        (../planning/01_GPS_Build_Reference.md §4.2) จึงไม่มีสถานะอุปกรณ์หลัง apply ใน response.
        โหมด mock ตรวจค่าเบื้องต้น (field ≥ 1, Timeout/Interval ไม่ติดลบ)
        แล้วคืน `applied: true`
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigApplyInput"
      responses:
        "200":
          description: "ผลการ apply (applied=false = ค่าใน Config ไม่ผ่านการตรวจเบื้องต้น)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigApplyResult"
        "400":
          description: configId ไม่ใช่ uuid ที่ถูกต้อง
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ใส่ Config ให้อุปกรณ์ — เฉพาะ Role ST, OT
        "404":
          description: ไม่พบ Device (deviceId) นี้ หรือไม่พบ Config (configId) นี้
        "409":
          description: >-
            ใส่ Config ไม่ได้ — Device ยังไม่ installed / Config ยังไม่ผ่านอนุมัติ
            (ต้อง approved หรือ synced) / Config คนละ deviceModel/protocol กับ Device
  /devices/{deviceId}/confirm-firmware-install:
    post:
      tags: [device]
      operationId: confirmFirmwareInstall
      summary: ช่างยืนยันว่าติดตั้ง Firmware เข้าอุปกรณ์เสร็จแล้วจริง (v3.29, issue #181)
      description: >-
        สำหรับช่างหน้างาน (ST/OT ผ่าน Mobile) — **ไม่ใช่คำสั่งส่งอะไรไปอุปกรณ์เลย**
        ต่างจาก `applyConfigToDevice` ที่ยังมี mock fire-and-forget อยู่เบื้องหลัง
        endpoint นี้เป็นแค่การบันทึกว่า "มีการยืนยันเกิดขึ้น" (AuditLog เท่านั้น)
        **ไม่คำนวณ `Firmware.deviceUpdateStatus` จริง** (ยังคงเป็น unknown ต่อไป)
        — placeholder ระดับที่ทำได้ในขอบเขตฝึกงาน ต่างจาก Workflow Firmware
        แบบ state machine อัตโนมัติเต็มรูปแบบใน design doc §10.2-10.3 ที่ทำไม่ได้
        เพราะระบบเป็น PULL model (backend ไม่คุยกับอุปกรณ์ตรงๆ)
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmFirmwareInstallInput"
      responses:
        "200":
          description: ยืนยันสำเร็จ — บันทึก AuditLog แล้ว
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmFirmwareInstallResult"
        "400":
          description: firmwareId ไม่ใช่ uuid ที่ถูกต้อง
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ยืนยันติดตั้ง Firmware — เฉพาะ Role ST, OT
        "404":
          description: ไม่พบ Device (deviceId) นี้ หรือไม่พบ Firmware (firmwareId) นี้
        "409":
          description: >-
            ยืนยันไม่ได้ — Device ยังไม่ installed / Firmware ยังไม่ผ่านอัปโหลด
            (ต้อง stored) / Firmware ยังไม่ผ่านอนุมัติคุณภาพ (ต้อง approved) /
            Firmware ไม่รองรับรุ่นอุปกรณ์เครื่องนี้ / **(v3.43, Sprint 3 แถวที่
            24)** firmwareId ที่ขอไม่ตรงกับ Campaign Rollout (Firmware) ที่
            `active` กำหนดไว้ให้อุปกรณ์เครื่องนี้ และไม่มี
            `DeviceFirmwareOverride` สถานะ `approved` รองรับ — ต้องขอ Firmware
            Override ก่อน (ดู `overrideDeviceFirmware`)
  /devices/{deviceId}/firmware-override:
    post:
      tags: [device]
      operationId: overrideDeviceFirmware
      summary: ST ขอให้อุปกรณ์เครื่องนี้ติดตั้ง Firmware ที่ระบุได้ แม้ไม่ตรงกับ Campaign Rollout ที่กำหนดไว้ (Firmware Override รายเครื่อง — Sprint 3 แถวที่ 24, v3.43)
      description: >-
        mirror `overrideDeviceConfig` เกือบทั้งหมด ต่างกันที่ไม่มี `fields`
        เพราะ Firmware เป็นเวอร์ชันเดียวทั้งก้อน ไม่ใช่ key-value ที่ override
        บางส่วนได้ — มีแค่ "อยากได้ firmware ตัวไหน" สร้างคำขอสถานะ `pending`
        เท่านั้น **ไม่มีผลทันที** ต้องรอ Operation อนุมัติผ่าน
        `approveDeviceFirmwareOverride` ก่อน (Separation of Duty) ถึงจะใช้
        `confirmFirmwareInstall` กับ firmware นี้ได้จริงถ้าขัดกับ Campaign
        Rollout ที่ `active` กำหนดไว้ (ดู 409 ใหม่ของ `confirmFirmwareInstall`)
        เครื่องหนึ่งมีคำขอ `pending` พร้อมกันได้แค่ 1 รายการ (409 ถ้ามีอยู่แล้ว)
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firmwareId, reason]
              properties:
                firmwareId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  maxLength: 500
                  description: เหตุผลของการขอ override — บังคับกรอกเสมอ
      responses:
        "200":
          description: >-
            ส่งคำขอสำเร็จ — คืนคำขอที่เพิ่งสร้าง (สถานะ `pending` เสมอ)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceFirmwareOverride"
        "400":
          description: ไม่ส่ง reason มา หรือ firmwareId ไม่ใช่ uuid ที่ถูกต้อง
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ override Firmware ของอุปกรณ์นี้ — เฉพาะ Role ST เท่านั้น (OT ไม่มีสิทธิ์เลย)
        "404":
          description: ไม่พบ Device (deviceId) นี้ หรือไม่พบ Firmware (firmwareId) นี้
        "409":
          description: >-
            Firmware ยังติดตั้งไม่ได้ (ยังไม่ stored/approved/รองรับรุ่นอุปกรณ์
            นี้) หรืออุปกรณ์นี้มีคำขอ override ที่รอ Operation อนุมัติอยู่แล้ว
...
  /devices/{deviceId}/simulate-config:
    post:
      tags: [device]
      operationId: simulateConfigOnDevice
      summary: เช็คว่า Config พร้อมอัพโหลดเข้าอุปกรณ์เครื่องนี้จริงไหม (v3.15)
      description: >-
        สำหรับช่างหน้างาน (ST/OT ผ่าน Mobile) — dry-run ก่อน `applyConfigToDevice`
        จริง รวม 3 ส่วนเป็นผลเดียว: (1) Config เองพร้อมไหม (reuse ตัวเดียวกับ
        `simulateConfig`) (2) deviceModel/protocol ของ Config ตรงกับอุปกรณ์เครื่อง
        นี้ไหม (3) สัญญาณอุปกรณ์จริง (reuse ตัวเดียวกับ `testDeviceConnection`).
        **ไม่ persist / ไม่แตะสถานะใดๆ** ทั้ง Device และ Config. อ้างด้วยเลขเครื่อง
        จริง (`Device.deviceId`). ต่างจาก `applyConfigToDevice`: deviceModel/protocol
        ไม่ตรง **ไม่ใช่ 409** — คืน 200 พร้อม `compatibilityCheck.passed:false`
        เพื่อให้ช่างเห็นว่าอะไรไม่ตรง
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateConfigOnDeviceInput"
      responses:
        "200":
          description: >-
            ผลเช็คความพร้อม (3 ส่วนรวมใน response เดียว) — `passed:false` = ส่วนใด
            ส่วนหนึ่งไม่ผ่าน (รวมถึง deviceModel/protocol ไม่ตรง)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSimulateConfigResult"
        "400":
          description: configId ไม่ใช่ uuid ที่ถูกต้อง
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์เช็คความพร้อมของอุปกรณ์ — เฉพาะ Role ST, OT
        "404":
          description: ไม่พบ Device (deviceId) นี้ หรือไม่พบ Config (configId) นี้
        "409":
          description: >-
            เช็คความพร้อมไม่ได้ — Device ยังไม่ installed หรือ Config ยังไม่ผ่าน
            อนุมัติ (ต้อง approved หรือ synced)
  /devices/{deviceId}/config:
    get:
      tags: [device]
      operationId: getDeviceCurrentConfig
      summary: Config ปัจจุบันของอุปกรณ์เครื่องนี้ (Config Override Phase 2, Mobile — issue #211, v3.32)
      description: >-
        derive จาก `Task.configId` ของ Task ล่าสุด (`status: completed`) ที่ผูก
        กับอุปกรณ์เครื่องนี้ — Device ไม่มี FK ตรงไป Config เลย (เหมือน
        `getDeviceStatus` ที่ยังไม่ implement) ใช้เปิดหน้า "Override ค่า
        พารามิเตอร์" ฝั่ง Mobile (ST เท่านั้น — resource `device-config-override`)
        ก่อนเรียก `overrideDeviceConfig` ต่อ · **v3.31 (issue #223):** ถ้า
        อุปกรณ์เครื่องนี้มี `DeviceConfigOverride` อยู่ ค่า `fields` ที่คืนกลับ
        เป็นค่าที่ merge override ทับ base Config แล้ว (ดู `hasDeviceOverride`)
        · **v3.32 (มติ 2026-09-24, PR #225):** merge เฉพาะแถวที่ Operation
        อนุมัติแล้ว (`status: approved`) เท่านั้น — คำขอที่ยัง `pending`/ถูก
        `rejected` ไม่มีผลกับ `fields`/`hasDeviceOverride` เลย ดูได้ผ่าน
        `pendingOverride` แยกต่างหาก (ถ้ามีคำขอค้างอยู่)
      parameters:
        - in: path
          name: deviceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config ปัจจุบันของอุปกรณ์ (merge override เฉพาะเครื่องแล้วถ้ามี)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceCurrentConfig"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์อ่าน Config ปัจจุบันของอุปกรณ์ — เฉพาะ Role ST, OT
        "404":
          description: >-
            ไม่พบ Device (deviceId) นี้ หรืออุปกรณ์นี้ยังไม่มี Config ที่ยืนยัน
            ติดตั้งแล้ว (ไม่มี Task สถานะ completed ที่ผูก configId กับอุปกรณ์นี้)
...
  /device-config-overrides:
    get:
      tags: [device]
      operationId: listDeviceConfigOverrides
      summary: Operation ดูรายการคำขอ Per-device Config Override ทั้งหมด (issue #223, มติ 2026-09-24, v3.32)
      description: >-
        คิวอนุมัติของ Operation — filter ตาม `status` (optional, ไม่ส่ง = ทุก
        สถานะ) เรียงคำขอใหม่สุดก่อน (`overriddenAt` desc) resource
        `device-config-override` action `Read` — Operation เท่านั้น
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: รายการคำขอ (เรียงตาม overriddenAt desc)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceConfigOverride"
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
        "403":
          description: ไม่มีสิทธิ์ดูคิวอนุมัติ override — เฉพาะ Role Operation เท่านั้น
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SimulationResult:
      type: object
      properties:
        passed:
          type: boolean
        details:
          type: array
          items:
            type: string
    DeviceConnectionTestResult:
      type: object
      description: ผลทดสอบการเชื่อมต่อของอุปกรณ์ที่ติดตั้งจริง (`testDeviceConnection`)
      required: [passed, signalStrength, details, testedAt]
      properties:
        passed:
          type: boolean
        signalStrength:
          type: integer
          description: ความแรงสัญญาณ (dBm) — mock mode คืนค่าคงที่ให้ Mobile ทำ/ทดสอบ UI ได้
          example: -65
        details:
          type: array
          items:
            type: string
        testedAt:
          type: string
          format: date-time
...
    RegisterDeviceRequest:
      type: object
      description: Request body สำหรับ registerDevice (issue #157 PR 1)
      required: [deviceId, simNumber, modelId, protocol]
      properties:
        deviceId:
          type: string
          maxLength: 50
          description: เลขเครื่องจริงที่จะลงทะเบียน (unique ทั้งระบบ)
          example: DEV-0099
        simNumber:
          type: string
          maxLength: 50
        modelId:
          type: string
          format: uuid
          description: >-
            id ของ DeviceModel ที่ลงทะเบียนไว้แล้ว (ดู GET /device-models) —
            `deviceModel` (string) ของ Device ที่สร้างจะ resolve จาก
            `model.name` เสมอ ไม่รับแยกจาก client
        protocol:
          type: string
          maxLength: 50
          description: ต้องอยู่ใน DeviceModel.supportedProtocols ของ modelId ที่ระบุ
          example: TCP
        hardwareRevisionCode:
          type: string
          maxLength: 50
          nullable: true
        customerId:
          type: string
          format: uuid
          nullable: true
    RegisterDeviceResponse:
      allOf:
        - $ref: "#/components/schemas/Device"
        - type: object
          required: [apiKey]
          properties:
            apiKey:
              type: string
              description: >-
                API key จริง (ไม่ hash) — โชว์ได้ครั้งเดียวตอนลงทะเบียนสำเร็จ
                เท่านั้น backend เก็บแค่ bcrypt hash ไว้ ไม่มีทาง GET
                กลับมาดูค่านี้ซ้ำได้อีก
...